feat: adding draft generation of react hook
