+++ package.json
@@ -32,6 +32,7 @@
     "yargs": "^17.1.1"
   },
   "devDependencies": {
+    "@types/react": "^17.0.21",
     "i18next": "^20.6.1",
     "simple-git-hooks": "^2.6.1"
   },
