+++ tsconfig.json
@@ -7,6 +7,7 @@
     "esModuleInterop": true,
     "downlevelIteration": true,
     "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true
+    "forceConsistentCasingInFileNames": true,
+    "jsx": "react"
   }
 }
